Changed float precision in gtsrb and stl10 datasets

+++ datasets/stl10.py
@@ -13,7 +13,7 @@ def _load_images(url):
         everything = np.fromfile(f, dtype=np.uint8)
         images = np.reshape(everything, (-1, 3, 96, 96))
         images = np.transpose(images, (0, 3, 2, 1))
-        images = np.array([rgb2gray(img) for img in images])
+        images = np.array([rgb2gray(img).astype(np.float32) for img in images])
         return images
